@@ -1,24 +1,23 @@
-namespace CommonNet.Extensions.Tests
+﻿using FsCheck;
+
+namespace CommonNet.Extensions.Tests;
+
+public static class NonNullNoCrLfStringArbitrary
 {
-    using FsCheck;
-
-    public static class NonNullNoCrlLfStringArbitrary
+    public static Arbitrary<string> Strings()
     {
-        public static Arbitrary<string> Strings()
-        {
-            return Arb.Default.String().Filter(x =>
-                    x != null &&
-                    !x.Contains("\n") &&
-                    !x.Contains("\r")
-                );
-        }
+        return Arb.Default.String().Filter(x =>
+                x != null &&
+                !x.Contains('\n') &&
+                !x.Contains('\r')
+            );
     }
+}
 
-    public static class NonEmptyArrayArbitrary
+public static class NonEmptyArrayArbitrary
+{
+    public static Arbitrary<T[]> Array<T>()
     {
-        public static Arbitrary<T[]> Array<T>()
-        {
-            return Arb.Default.Array<T>().Filter(x => x != null && x.Length > 0);
-        }
+        return Arb.Default.Array<T>().Filter(x => x != null && x.Length > 0);
     }
 }
